Fixed initialisation and increased speed
I think I have to leave README.md from being my calendar and at last
track changes with 'commmit' messages. So, the first thing I took care
of these days was to remake the initialisation process, which made the
algorithm hand in every run. I created a method with which previous
segmentation data is compared with current setup and, if no serious
changes exist, segmentation data is loaded and segmentation is skipped.
Apart from that, I brushed up a little the code, threw unnecessary
modules and function calls  that lagged it and  moved the necessary
constants for tweaking to the config.yaml file, allowing to everyone
(with the correct setup) to run it. I am continuing my little struggle
with constants tweaking, while on the hunt for a good way for the next
step. Apparently the new README file will also be prepared soon.