Completed feature extraction part
I completed the part where the ROI features are gathered. Now follows
the sparse coding implementation and the dictionary training code. I
think I will add one more file as the class_objects.py has more than 850
lines now and tends to get unreadable, so an action_recognition_alg.py
is imminent. I also need to find some dataset to start with, as there is
no possible way to test what I have done till now, concerning the
fanello13a algorithm. It can be one-shot, but it can also be vaster,
just to reach a local maximum of the technique. Major drawback is the
fact that no dataset exists in the internet, as far as I can tell,that
concerns depth data of hand actions with the camera facing the hand from the back.
If anyone has found anything, it would be awesome to share and his name
will get to be in my Thesis as a contributor or at least in the 'Thanks'
section.Apart from that, I need to create a figure of the setup so that anyone
can understand what I am doing, in order to be able to help somehow.