Completed sparse coding
*Sparse coding dictionary learning should be run many times, so that to reassure convergence and error minimization, as fmin_tnc might not run well, due to bad initialisation (still cannot understand the reason for this)*
Next step is to get my hands on some dataset, so that I can modify them
in the way the paper describes. After that, I will pass them as input to
the sparse coding algorithm, training the bases and finding their
inverse. Next stage will be to find a good environment and mine some
data for training.