Put the algorithm together
Using 'pinch' data from Dexter 1 dataset, I tested the current set up.
After some modifications in the sparse coding stage and aleviation of
some not so obvious problems, I managed to get it all up and working and
extract the 3DHOF and GHOG dictionaries.
Now, I need to make things as instructed from the rest of the paper,
that is get batches of video data as input and train the svm linear
classifiers by using the sparsity dictionaries.