Added hist4d visualization
After 5 days of vacations, I began testing the method described in the
paper with Dexter 1 data. The results were pretty disappointing,
therefore I assume there must be a bug either in the feature extraction
process (can be easily fixed) or bad understanding of the algorithm (unprobable). To this end, I have created a module that can visualize 4d histograms in space, in order to check the 3dHOFs. The GHOG features are more easily tested. The debugging continues...