Added data farming interface
Having no other way to test what I have done till now, I am forced to
try on some personal datasets. Thus, I created a GUI program, that will
receive a recorded rosbag, extract the frames and allow to a user to
find the frames indices that contain a specific action. The indices,
along with the actions names, are then saved in a csv file. The next
thing to do is to receive this csv and the recorded rosbag inside
init_code, extract the hand masks and create either a corresponding
directory which will include them, or make a list on the fly and save it
inside pkl file. Using the first way might be better, but more
polluting. Either way, after masks are computed, we will register them to
the action recognition algorithm, extract the features, train
the svms and test the created machine. ~1000 lines of code away from
completion, if everything goes well.