Major Changes towards optimization
Action Recognition: Transformed current classifier to an alpha toolbox, which
has the ability to get as input a stream and test it based on a trained model.
In order to test this, I needed some good training data. I did the following.

Data Mining : I added a way to test the MOG2 parameters through gui, which got
friendlier overall to the user. I recorded a rosbag including 9 actions, 3 from
which involved only passive gesture, moving randomly in space. I saw that mog2
have some issues when the hand remains static for long time inside the field of
view, causing part of the hand to be considered background.

New algorithm in the game: Due to this behavior, I implemented with the help of
opencv a depth oriented camshift algorithm, which will be used to fast track
the hand based on depth , with input a memory of hand masks. Apart from this
upgrade, I aim to recreate the hand segmentation algorithm, however the main axis
will remain the same.

Hand Segmentation Algorithm: In order for the following to happen, I changed
the way entry to the image is found, adding support for multiple hands, which
could be manipulated through multiprocessing (haven't gotten to that yet) and I
now give a patch of the image to the main process, not the whole one. So the
new algorithm will retain the way the joints and links of the hand are found,
but the abnormality concept will be replaced for something safer to work with,
as this concept needs so many parameters to be fixed and it is not in any way
invariable to hand shape. The replace will be an overlapping patched oriented
search, based on the found links, from which hog will be extracted. I strongly
believe, based on papers and readings, that hog features are the best way to
extract the curvature information from the arm. Each histogram will have a
variance between its bins, which will represent its score. This will be
because hand is the part of the arm with the highest curvature variance in
every dimension. I will create a score map and using non-maxima suppression I will extract the hand shape and
location. This will give me better results.

The camshift and the new hand segmentation algorithm might offer me to present
state of the art results.

Unfortunately, I believe that the deadline for the end of February can not be
aimed anymore, but I still hope.