Enabled online testing, needs kinect support
I have enabled online testing in the SVM implementation. The algorithm
is just like fanello had proposed (they should really at least feel
honored for my blind trust to them, the absence of their code was surely
not a promising thing, but did bet on them and they seem to actually
fullfill my expectations) . I can get above 90% accuracy in
easily discriminable gestures (punch, pinch, palm, tiger grasp etc.) but
actions that are orientationally altering along time can not be so easily
discriminated (eg. hand up, hand down), which might be corrected if I
add samples to the training dataset. However, due to the necessity that
the hand axis should remain stable while performing these actions, the
3DHOF features are not trained well under any circumstance. Well, the
almost realtime 90% accuracy (must be around 10 fps with only one CPU
and no multithreading, haven't checked the whole pipeline yet, I need a kinect to be joined to the current project
and this is like 100 lines of code, which will be there in the next
update) with the other gestures seems promising enough. So, the next
step is to create the framework where this project will run and to
specialize the application, as far as I can, to be able to draw
something! If I can get to this critical point, then the algorithm is
almost ready for shipping (partyyy). Some things to link to gestures are:
-Enable drawing using the index finger (easy, as I have the principal
axis of the hand, which passes right through the index when it is
straight most times)
-Enable erasing with the punch (a moving disk/square sweeping everything,
yaaay)
-Enable tiger grasp to drag. This is more like a programmer's fun.
-Enable zooming with the pinch. Well this is hard, but I really want to
put my hands on it and make some small progress there. I just need to
somehow calculate the distance between thumb and index, even when they
are kinda joined. The zooming process is something easily programmable.

If I am serious enough to proceed following my initial targets, a time
profiling of the code is demanded, a cythonizing of some of it, a cpping
also of some opencv methods (damn, I wished it had gpu wrappers for
python, my Cpp knowledge is bounded by C, have to deal with
some alien coding now, hehe...) and generally a restyling of my code,
which could be combinated with writing my Thesis (I wish I could just
send my ~7000 lines of code and say 'look it is all here!'), which is
designated to become a latex project (~ 1 month full work, maybe more,
have to be happy with this deadline ->middle April???)