Activated online testing and deduced a combined classifier
After experimentation, I saw that the computational weight of sparse
coding was just not worth it, so I changed my thinking and now use
classifiers with pure 3DXYPCA, 3DHOF and GHOG. I joined the two
classifiers, hoping for the bettter, using an estimator deduced by the
bayesian logic. Till now, I have gotten a 96% in offline testing, but
where I thought this combination would excel was the online one, which
however gets me only 86% maximum which is pretty frustrating. I will try
to fix this. In other news, I allowed the both classifiers to run
realtime (hurray), but the dataset, particularly the dynamic actions
one, is too small. I am thinking of also adding a gui application for
adding and removing actions, in order to be able to keep record of the
present actions and configuration.