Bugfixes and complete modification of results organisation
In search of better results, I broke my code, thus had to debug it.
Well, this took time, but I think that I can produce correct results
now. The worst I forgot was to implement the Farneback algorithm, as
proposed by Fanello. Unfortunately, due to the non uniform nature of
depth intensity in time, any optical flow in depth space seems like a
big mistake. So, the 3DHOF features are out of the game, but will be
included in the results for completeness (not being rubbish and having
worked a lot on them, plus that I have upgrade them a little to be
invariant to link rotation, using my simple skeletonization process).
Apart from that, due to the many variations one can run the
classification stage, I decided to keep a dictionary of classifiers
saved in a simple single file, another dictionary to keep sparse coders
and another one to keep scores. The scores folders are now less deep and
their numbering confomrs to an autogenerated file inside the running
directory. Only drawback is the fact that when using SVC and not
LinearSVC the saved classifier is like 400MB+, 100 times more of the one
when using LinearSVC. I guess I might have to find a way to reduce this
size, if I decide to use an rbf kernel in the future.